fix: address CR review on v3.1.0
- lib.php: fix vendor/autoload.php path (../vendor → ./vendor)
- lib.php: capture_utilization_snapshot skips subnets already snapped today
- api.php: expiring_days uses exclusive upper bound (< not <=)
- api.php: api_utilization_snapshots returns 404 for unknown subnet_id
- addresses.php: carry filter/days params through sort and pagination URLs
- assets/app.js: move SMTP test button handler out of inline script (CSP fix)
- assets/app.css: add expired-row dark override for prefers-color-scheme auto
- settings.php: remove inline script that violated CSP script-src 'self'
- export_utilization_history.php: require_role('admin') to match reports.php
- schema.pgsql.sql: use BIGINT for utilization_snapshots.subnet_id FK
- tests: update MysqlSmokeTest + PgsqlSmokeTest migration count 33 → 36
- README.md: add MariaDB to requirements table

+++ Simple-PHP-IPAM/schema.pgsql.sql
@@ -400,7 +400,7 @@ CREATE TABLE IF NOT EXISTS alert_state (
 -- ---------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS utilization_snapshots (
   id          SERIAL PRIMARY KEY,
-  subnet_id   INTEGER NOT NULL REFERENCES subnets(id) ON DELETE CASCADE,
+  subnet_id   BIGINT NOT NULL REFERENCES subnets(id) ON DELETE CASCADE,
   snapped_at  TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   used_count  INTEGER NOT NULL,
   free_count  INTEGER NOT NULL,
